python_ai/requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
keyboard==0.13.5
pyautogui==0.9.54
torch==2.0.1
gymnasium==0.29.1 